Fix authentication with query strings
When using a query string with a GET request (for instance when
doing a search), authentication breaks because the query string is
included in the input for the Hashed Path authentication header.

Per the documentation for the Chef API, query strings are not
supposed to be included.

  cf. https://docs.chef.io/api_chef_server.html#header_format

Replacing the call to url#path to url#pathname fixes the issue.